+++ bauh/gems/arch/confirmation.py
@@ -16,7 +16,7 @@ def request_optional_deps(pkgname: str, pkg_mirrors: dict, watcher: ProcessWatch
     opts = []
 
     for p, d in pkg_mirrors.items():
-        op = InputOption('{}{} ( {}: {} )'.format(p, ': ' + d['desc'] if d['desc'] else '', i18n['mirror'], d['mirror'].upper()), p)
+        op = InputOption('{}{} ( {}: {} )'.format(p, ': ' + d['desc'] if d['desc'] else '', i18n['repository'], d['mirror'].upper()), p)
         op.icon_path = _get_mirror_icon(d['mirror'])
         opts.append(op)
 
@@ -39,7 +39,7 @@ def request_install_missing_deps(pkgname: str, pkg_mirrors: dict, watcher: Proce
 
     opts = []
     for p, m in pkg_mirrors.items():
-        op = InputOption('{} ( {}: {} )'.format(p, i18n['mirror'], m.upper()), p)
+        op = InputOption('{} ( {}: {} )'.format(p, i18n['repository'], m.upper()), p)
         op.read_only = True
         op.icon_path = _get_mirror_icon(m)
         opts.append(op)